floats changed to doubles

+++ stepDetector/src/main/java/com/keepfit/stepdetection/accelerometer/filter/Util.java
@@ -2,20 +2,20 @@ package com.keepfit.stepdetection.accelerometer.filter;
 
 public class Util {
 
-    public static double magnitude(float[] values) {
+    public static double magnitude(double[] values) {
         if (values == null) throw new IllegalArgumentException("ERROR: values arg cannot be null");
         if (values.length < 1)
             throw new IllegalArgumentException("ERROR: values array arg cannot be empty");
         return Math.sqrt(sumOfSquares(values));
     }
 
-    public static float sumOfSquares(float[] values) {
+    public static float sumOfSquares(double[] values) {
         if (values == null) throw new IllegalArgumentException("ERROR: values arg cannot be null");
         if (values.length < 1)
             throw new IllegalArgumentException("ERROR: values array arg cannot be empty");
         float sos = 0;
         double squared = 2;
-        for (float f : values) {
+        for (double f : values) {
             sos += Math.pow(f, squared);
         }
         return sos;
